requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
colorama==0.4.3
Django==3.1.1
django-bootstrap4==2.3.0
django-camera-imagefield==0.2
django-crispy-forms==1.9.2
django-filter==2.4.0
django-materializecss-form==1.1.17
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pylint==2.6.0
python-datauri==0.2.9
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
uuid==1.30
wrapt==1.12.1
